Cambios

+++ app/Http/Controllers/ProjectController.php
@@ -4,8 +4,12 @@ namespace App\Http\Controllers;
 
 
 use App\Models\Project;
-use App\Http\Requests\SaveProjectRequest;
 use Illuminate\Http\Request;
+use Intervention\Image\ImageManager;
+use Intervention\Image\Drivers\Gd\Driver;
+use Illuminate\Support\Facades\Storage;
+use App\Http\Requests\SaveProjectRequest;
+use App\Models\Category;
 
 class ProjectController extends Controller
 {
@@ -21,7 +25,9 @@ class ProjectController extends Controller
     {
 
         return view('projects.index', [
-            'projects' => Project::latest()->paginate()
+            'newProject' => new Project,
+            'projects' => Project::with('category')->latest()->paginate(),
+            'deletedProjects' => Project::onlyTrashed()->get()
         ]);
         
     }
@@ -38,8 +44,11 @@ class ProjectController extends Controller
 
     public function create()
     {
+        $this->authorize('create', $project = new Project);
+
         return view('projects.create', [
-            'project' => new Project
+            'project' => $project,
+            'categories' => Category::pluck('name', 'id')
         ]);
     }
 
@@ -47,29 +56,107 @@ class ProjectController extends Controller
     public function store(SaveProjectRequest $request)
     {
         
-        Project::create( $request->all() );
+        
+        $project = new Project( $request->validated() );
+
+        $this->authorize('create', $project);
+
+        $project->image = $request->file('image')->store('images','public');
+
+        $project->save();
+
+        $manager = new ImageManager(new Driver());
+
+        $image = $manager->read(storage_path('app/public/' . $project->image))->scale(width: 300)->reduceColors(16);
+        $encoded = $image->toJpeg(60);
+
+        Storage::put($project->image, (string) $encoded);
+
 
         return redirect()->route('projects.index');
     }
 
     public function edit (Project $project){
 
+        $this->authorize('update', $project);
+
         return view('projects.edit', [
-            'project' => $project
+            'project' => $project,
+            'categories' => Category::pluck('name', 'id')
         ]);
     }
 
     public function update(Project $project, SaveProjectRequest $request){
 
-        $project->update( $request->validated());
+        $this->authorize('update', $project);
+
+        if ( $request->hasFile('image') ){
+
+            
+            $project->fill( $request->validated() );
+
+            $project->image = $request->file('image')->store('images','public');
+
+            $project->save();
+
+
+            $manager = new ImageManager(new Driver());
+
+            $image = $manager->read(storage_path('app/public/' . $project->image))->scale(width: 300)->reduceColors(16);
+            $encoded = $image->toJpeg(60);
+
+            Storage::put($project->image, (string) $encoded);
+
+        }
+
+        else{
+
+            $project->update( array_filter( $request->validated()) );
+        }
+
+       
 
         return redirect()->route('projects.show', $project);
     }
 
     public function destroy(Project $project){
 
+        $this->authorize('delete', $project);
+
+
         $project->delete();
 
         return redirect()->route('projects.index');
     }
+
+    public function restore($projectUrl){
+
+        $project = Project::withTrashed()->whereUrl($projectUrl)->firstOrFail();
+
+        $this->authorize('restore', $project);
+
+        $project->restore();
+
+        return redirect()->route('projects.index');
+    }
+
+
+    public function ForceDelete($projectUrl)
+    {
+        $project = Project::withTrashed()->whereUrl($projectUrl)->firstOrFail();
+
+        $this->authorize('forceDelete', $project);
+
+
+        Storage::delete($project->image);
+
+        
+        $project->forceDelete();
+
+        return redirect()->route('projects.index');
+    }
+
+
+
+
 }
